refactor(frontend): break open-crypto-pay import cycle per Copilot review
